backend/requirements.txt
# Django Core
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Authentication
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# API Documentation
drf-spectacular==0.27.0

# File Storage
boto3==1.34.19
django-storages==1.14.2

# Utilities
python-dateutil==2.8.2
Pillow==10.1.0
django-filter==23.5

# Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
faker==22.0.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0
